環境に関するコードを整理した
