+++ elisp/src/env_single.rs
@@ -5,15 +5,14 @@
    hidekuno@gmail.com
 */
 use std::cell::RefCell;
-use std::collections::BTreeMap;
 use std::rc::Rc;
 use std::vec::Vec;
 
-use crate::buildin::{create_function, BuildInTable};
+use crate::env::{GlobalTbl, SimpleEnv};
 use crate::lisp::{BasicBuiltIn, Expression, Function, ResultExpression};
 //========================================================================
-type ExtFunction = dyn Fn(&[Expression], &Environment) -> ResultExpression;
-type EnvTable = Rc<RefCell<SimpleEnv>>;
+pub(crate) type ExtFunction = dyn Fn(&[Expression], &Environment) -> ResultExpression;
+pub(crate) type EnvTable = Rc<RefCell<SimpleEnv>>;
 //------------------------------------------------------------------------
 pub type FunctionRc = Rc<Function>;
 pub type ExtFunctionRc = Rc<ExtFunction>;
@@ -27,6 +26,20 @@ macro_rules! referlence_list {
 }
 #[macro_export]
 macro_rules! mut_list {
+    ($e: expr) => {
+        $e.borrow_mut();
+    };
+}
+
+#[macro_export]
+macro_rules! referlence_env {
+    ($e: expr) => {
+        $e.borrow();
+    };
+}
+
+#[macro_export]
+macro_rules! mut_env {
     ($e: expr) => {
         $e.borrow_mut();
     };
@@ -145,70 +158,5 @@ impl Environment {
             s.push_str("\n");
         }
         s
-    }
-}
-impl BuildInTable for BTreeMap<&'static str, BasicBuiltIn> {
-    fn regist(&mut self, symbol: &'static str, func: BasicBuiltIn) {
-        self.insert(symbol, func);
-    }
-}
-struct GlobalTbl {
-    builtin_tbl: BTreeMap<&'static str, BasicBuiltIn>,
-    builtin_tbl_ext: BTreeMap<&'static str, ExtFunctionRc>,
-    tail_recursion: bool,
-    force_stop: bool,
-}
-impl GlobalTbl {
-    pub fn new() -> Self {
-        let mut b: BTreeMap<&'static str, BasicBuiltIn> = BTreeMap::new();
-        create_function(&mut b);
-        GlobalTbl {
-            builtin_tbl: b,
-            builtin_tbl_ext: BTreeMap::new(),
-            tail_recursion: true,
-            force_stop: false,
-        }
-    }
-}
-pub struct SimpleEnv {
-    env_tbl: BTreeMap<String, Expression>,
-    parent: Option<EnvTable>,
-}
-impl SimpleEnv {
-    pub fn new(parent: Option<EnvTable>) -> Self {
-        if let Some(p) = parent {
-            SimpleEnv {
-                env_tbl: BTreeMap::new(),
-                parent: Some(p.clone()),
-            }
-        } else {
-            SimpleEnv {
-                env_tbl: BTreeMap::new(),
-                parent: parent,
-            }
-        }
-    }
-    pub fn regist(&mut self, key: String, exp: Expression) {
-        self.env_tbl.insert(key, exp);
-    }
-    pub fn find(&self, key: &String) -> Option<Expression> {
-        match self.env_tbl.get(key) {
-            Some(v) => Some(v.clone()),
-            None => match self.parent {
-                // p.borrow().find(key), cannot return value referencing temporary value
-                Some(ref p) => p.borrow().find(key),
-                None => None,
-            },
-        }
-    }
-    pub fn update(&mut self, key: &String, exp: Expression) {
-        if self.env_tbl.contains_key(key) {
-            self.env_tbl.insert(key.to_string(), exp);
-        } else {
-            match self.parent {
-                Some(ref p) => p.borrow_mut().update(key, exp),
-                None => {}
-            }
-        }
     }
 }
