+++ elisp/src/env_thread.rs
@@ -4,19 +4,17 @@
 
    hidekuno@gmail.com
 */
-
-use std::collections::HashMap;
 use std::sync::Arc;
 use std::sync::Mutex;
 use std::sync::RwLock;
 use std::vec::Vec;
 
-use crate::buildin::{create_function, BuildInTable};
+use crate::env::{GlobalTbl, SimpleEnv};
 use crate::lisp::{BasicBuiltIn, Expression, Function, ResultExpression};
 //========================================================================
-type ExtFunction =
+pub(crate) type ExtFunction =
     Box<dyn Fn(&[Expression], &Environment) -> ResultExpression + Sync + Send + 'static>;
-type EnvTable = Arc<Mutex<SimpleEnv>>;
+pub(crate) type EnvTable = Arc<Mutex<SimpleEnv>>;
 //------------------------------------------------------------------------
 pub type FunctionRc = Arc<Function>;
 pub type ExtFunctionRc = Arc<ExtFunction>;
@@ -46,6 +44,21 @@ macro_rules! mut_list {
         $e.write().unwrap()
     };
 }
+
+#[macro_export]
+macro_rules! referlence_env {
+    ($e: expr) => {
+        $e.lock().unwrap();
+    };
+}
+
+#[macro_export]
+macro_rules! mut_env {
+    ($e: expr) => {
+        $e.lock().unwrap();
+    };
+}
+
 #[derive(Clone)]
 pub struct Environment {
     core: EnvTable,
@@ -115,70 +128,5 @@ impl Environment {
     }
     pub fn is_force_stop(&self) -> bool {
         self.globals.lock().unwrap().force_stop
-    }
-}
-impl BuildInTable for HashMap<&'static str, BasicBuiltIn> {
-    fn regist(&mut self, symbol: &'static str, func: BasicBuiltIn) {
-        self.insert(symbol, func);
-    }
-}
-struct GlobalTbl {
-    builtin_tbl: HashMap<&'static str, BasicBuiltIn>,
-    builtin_tbl_ext: HashMap<&'static str, ExtFunctionRc>,
-    tail_recursion: bool,
-    force_stop: bool,
-}
-impl GlobalTbl {
-    fn new() -> Self {
-        let mut b: HashMap<&'static str, BasicBuiltIn> = HashMap::new();
-        create_function::<HashMap<&'static str, BasicBuiltIn>>(&mut b);
-        GlobalTbl {
-            builtin_tbl: b,
-            builtin_tbl_ext: HashMap::new(),
-            tail_recursion: true,
-            force_stop: false,
-        }
-    }
-}
-pub struct SimpleEnv {
-    env_tbl: HashMap<String, Expression>,
-    parent: Option<EnvTable>,
-}
-impl SimpleEnv {
-    fn new(parent: Option<EnvTable>) -> Self {
-        if let Some(p) = parent {
-            SimpleEnv {
-                env_tbl: HashMap::new(),
-                parent: Some(p.clone()),
-            }
-        } else {
-            SimpleEnv {
-                env_tbl: HashMap::new(),
-                parent: parent,
-            }
-        }
-    }
-    fn regist(&mut self, key: String, exp: Expression) {
-        self.env_tbl.insert(key, exp);
-    }
-    fn find(&self, key: &String) -> Option<Expression> {
-        match self.env_tbl.get(key) {
-            Some(v) => Some(v.clone()),
-            None => match self.parent {
-                // p.borrow().find(key), cannot return value referencing temporary value
-                Some(ref p) => p.lock().unwrap().find(key),
-                None => None,
-            },
-        }
-    }
-    fn update(&mut self, key: &String, exp: Expression) {
-        if self.env_tbl.contains_key(key) {
-            self.env_tbl.insert(key.to_string(), exp);
-        } else {
-            match self.parent {
-                Some(ref p) => p.lock().unwrap().update(key, exp),
-                None => {}
-            }
-        }
     }
 }
