+++ elisp/src/lib.rs
@@ -9,6 +9,7 @@
 extern crate lazy_static;
 pub mod buildin;
 pub mod chars;
+pub mod env;
 pub mod io;
 pub mod lisp;
 pub mod list;
